+++ Example/TypographyKit/AppDelegate.swift
@@ -7,6 +7,7 @@
 //
 
 import UIKit
+import TypographyKit
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
@@ -15,6 +16,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
     internal func application(_ application: UIApplication, didFinishLaunchingWithOptions
         launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        TypographyKit.configurationURL = Bundle.main.url(forResource: "TypographyKit", withExtension: "json")
         return true
     }