+++ TypographyKit/Classes/ConfigurationKey.swift
@@ -13,5 +13,6 @@ enum ConfigurationKey: String {
     case maximumPointSize = "maximum-point-size"
     case minimumPointSize = "minimum-point-size"
     case pointSize = "point-size"
+    case scalingMode = "scaling-mode"
     case textColor = "text-color"
 }